(1) Added call from ImageIndex to FeatureFactory for getting image features to index (2) Changed FeatureFactory to use only first three image bands (3) Changed image features from list of doubles to list of integers based on what is returned from FeatureFactory

+++ lsh/indexing/HashFunction.java
@@ -48,7 +48,7 @@ public class HashFunction implements Serializable {
 	 * @param objectFeatures
 	 * @return slot number for locality sensitive hashing
 	 */
-	public int getSlotNumber(List<Double> objectFeatures) {
+	public int getSlotNumber(List<Integer> objectFeatures) {
 		
 		assert objectFeatures.size() == this.hashFunctionCoefficients.size() : 
 			"Size mismatch between object to be hashed and hash function";
